Added broken code
What is the error?
There is no error, it practically does what it has to do, but when it exits back to the original "OS", the shell somehow doesn't work anymore and errors out.
How to reproduce?
```
vmcore.lua insertBios /testhd /testbios
vmcore.lua runHD /testhd
```

I'm completely restarting the project now.
If you have a fix for this problem, please make a PULL-Request